pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fairmop"
version = "0.1.0"
description = "FairMOP: Benchmarking Fairness-Utility Trade-offs in Text-to-Image Models via Pareto Frontiers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "FairMOP Team"},
]
keywords = [
    "fairness",
    "text-to-image",
    "benchmarking",
    "pareto",
    "multi-objective optimization",
    "generative AI",
    "bias",
    "VLM",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "numpy>=1.24.0",
    "Pillow>=9.0.0",
    "tqdm>=4.64.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
clip = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    # CLIP must be installed separately:
    # pip install git+https://github.com/openai/CLIP.git
]
fid = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "clean-fid>=0.1.35",
    "scipy>=1.10.0",
]
prdc = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "prdc>=0.2",
    "scipy>=1.10.0",
]
datasets = [
    "huggingface_hub>=0.20.0",
]
openai = [
    "openai>=1.0.0",
]
gemini = [
    "google-genai>=0.5.0",
]
dashboard = [
    "streamlit>=1.28.0",
    "plotly>=5.15.0",
    "pandas>=2.0.0",
    "kaleido>=0.2.1",
]
all = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "clean-fid>=0.1.35",
    "prdc>=0.2",
    "scipy>=1.10.0",
    "openai>=1.0.0",
    "google-genai>=0.5.0",
    "streamlit>=1.28.0",
    "plotly>=5.15.0",
    "pandas>=2.0.0",
    "kaleido>=0.2.1",
    "huggingface_hub>=0.20.0",
]

[project.urls]
Homepage = "https://github.com/Malta-Lab/FairMOP"
Repository = "https://github.com/Malta-Lab/FairMOP"

[project.scripts]
fairmop = "fairmop.__main__:main"

[tool.setuptools.packages.find]
include = ["fairmop*"]
exclude = ["tests*", "look_here*"]
